filter product

+++ BackEnd_SpringBoot/src/main/java/com/trinhminhthaito/backend_springboot/services/Imp/OrderServicesImp.java
@@ -3,10 +3,15 @@ package com.trinhminhthaito.backend_springboot.services.Imp;
 import com.trinhminhthaito.backend_springboot.dtos.response.MessageDataResponse;
 import com.trinhminhthaito.backend_springboot.dtos.response.MessageResponse;
 import com.trinhminhthaito.backend_springboot.dtos.response.OrderListUserRepository;
+import com.trinhminhthaito.backend_springboot.models.orderModels.ItemsOrder;
 import com.trinhminhthaito.backend_springboot.models.orderModels.Order;
 import com.trinhminhthaito.backend_springboot.models.orderModels.PaymentDetail;
+import com.trinhminhthaito.backend_springboot.models.productModels.Product;
 import com.trinhminhthaito.backend_springboot.repository.OrderRepository;
+import com.trinhminhthaito.backend_springboot.repository.ProductRepository;
 import com.trinhminhthaito.backend_springboot.services.OrderServices;
+
+import jakarta.mail.FetchProfile.Item;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,10 +21,12 @@ import java.util.*;
 public class OrderServicesImp implements OrderServices {
 
 	private final OrderRepository orderRepository;
+	private final ProductRepository productRepository;
 
 	@Autowired
-	public OrderServicesImp(OrderRepository orderRepository) {
+	public OrderServicesImp(OrderRepository orderRepository, ProductRepository productRepository) {
 		this.orderRepository = orderRepository;
+		this.productRepository = productRepository;
 	}
 
 	// fn: hàm tạo mã đơn hàng
@@ -35,12 +42,63 @@ public class OrderServicesImp implements OrderServices {
 		return orderId;
 	}
 
+	// fn: hàm giảm số lượng hàng tồn database
+	private MessageResponse reduceStockProduct(List<ItemsOrder> listOrder) {
+		MessageResponse messageResponse = new MessageResponse();
+		try {
+			for (ItemsOrder item : listOrder) {
+				Optional<Product> productOption = productRepository.findById(item.getProductId());
+				Product product = productOption.get();
+				if (product != null) {
+					int stockNew = product.getStock() - (int) item.getStock();
+					product.setStock(stockNew);
+					messageResponse.setCode(0);
+				} else {
+					messageResponse.setCode(1);
+					messageResponse.setMessage("Không còn sản phẩm");
+				}
+			}
+		} catch (Exception ex) {
+			messageResponse.setCode(-1);
+			messageResponse.setMessage("Lỗi server: " + ex.toString());
+		}
+		return messageResponse;
+	}
+
+	// fn: hàm cộng lại số lượng hàng tồn khi huỷ đơn hàng
+	private MessageResponse addStockProduct(List<ItemsOrder> list) {
+		MessageResponse messageResponse = new MessageResponse();
+		try {
+			for (ItemsOrder item : list) {
+				Optional<Product> productOption = productRepository.findById(item.getProductId());
+				Product product = productOption.get();
+				if (product != null) {
+					int stockNew = product.getStock() + (int) item.getStock();
+					product.setStock(stockNew);
+					messageResponse.setCode(0);
+				} else {
+					messageResponse.setCode(1);
+					messageResponse.setMessage("Không còn sản phẩm");
+				}
+			}
+		} catch (Exception ex) {
+			messageResponse.setCode(-1);
+			messageResponse.setMessage("Lỗi server: " + ex.toString());
+		}
+		return messageResponse;
+	}
+
 	// fn: tạo đơn hàng
 	@Override
 	public MessageResponse createOrder(Order order) {
 		MessageResponse messageResponse = new MessageResponse();
 		try {
 			String orderId = createOrderId();
+
+			messageResponse = reduceStockProduct(order.getItemsOrders());
+			if (messageResponse.getCode() != 0)
+				return messageResponse;
+
 			order.setOrderCode(orderId);
 			orderRepository.save(order);
 			messageResponse.setCode(0);
@@ -50,37 +108,6 @@ public class OrderServicesImp implements OrderServices {
 			messageResponse.setMessage("Lỗi server: " + ex.getMessage());
 		}
 		return messageResponse;
-	}
-
-	// fn: get danh sách đơn hàng theo id người dùng
-	@Override
-	public MessageDataResponse getOrderById(String customerId) {
-		MessageDataResponse messageDataResponse = new MessageDataResponse();
-		try {
-			List<Order> listOrder = orderRepository.findByCustomerOrderCustomerId(customerId);
-			List<OrderListUserRepository> orderList = new ArrayList<>();
-			if (listOrder.isEmpty()) {
-				messageDataResponse.setCode(1); // Custom code for no data found
-				messageDataResponse.setMessage("No orders found for customerId: " + customerId);
-			} else {
-				for (Order item : listOrder) {
-					OrderListUserRepository o = new OrderListUserRepository();
-					PaymentDetail pd = item.getPaymentDetail();
-					o.setOrderCode(item.getOrderCode());
-					o.setOrderDate(item.getOrderDate());
-					o.setStatus(item.getStatus());
-					o.setPaidAmount(pd.getPaidAmount());
-					orderList.add(o);
-				}
-				messageDataResponse.setCode(0); // Custom code for success
-				messageDataResponse.setMessage("Success");
-				messageDataResponse.setData(orderList); // Set the list of orders as data
-			}
-		} catch (Exception ex) {
-			messageDataResponse.setCode(-1); // Custom code for error
-			messageDataResponse.setMessage("Server error: " + ex.getMessage());
-		}
-		return messageDataResponse;
 	}
 
 	// fn: update payment status
@@ -122,6 +149,10 @@ public class OrderServicesImp implements OrderServices {
 				messageResponse.setCode(0);
 				messageResponse.setMessage("Suscess");
 			} else {
+				messageResponse = addStockProduct(order.getItemsOrders());
+				if (messageResponse.getCode() != 0)
+					return messageResponse;
+
 				orderRepository.delete(order);
 				messageResponse.setCode(0);
 				messageResponse.setMessage("Suscess");
@@ -131,6 +162,37 @@ public class OrderServicesImp implements OrderServices {
 			messageResponse.setMessage("Lỗi server: " + ex.getMessage());
 		}
 		return messageResponse;
+	}
+
+	// fn: get danh sách đơn hàng theo id người dùng
+	@Override
+	public MessageDataResponse getOrderById(String customerId) {
+		MessageDataResponse messageDataResponse = new MessageDataResponse();
+		try {
+			List<Order> listOrder = orderRepository.findByCustomerOrderCustomerId(customerId);
+			List<OrderListUserRepository> orderList = new ArrayList<>();
+			if (listOrder.isEmpty()) {
+				messageDataResponse.setCode(1); // Custom code for no data found
+				messageDataResponse.setMessage("No orders found for customerId: " + customerId);
+			} else {
+				for (Order item : listOrder) {
+					OrderListUserRepository o = new OrderListUserRepository();
+					PaymentDetail pd = item.getPaymentDetail();
+					o.setOrderCode(item.getOrderCode());
+					o.setOrderDate(item.getOrderDate());
+					o.setStatus(item.getStatus());
+					o.setPaidAmount(pd.getPaidAmount());
+					orderList.add(o);
+				}
+				messageDataResponse.setCode(0); // Custom code for success
+				messageDataResponse.setMessage("Success");
+				messageDataResponse.setData(orderList); // Set the list of orders as data
+			}
+		} catch (Exception ex) {
+			messageDataResponse.setCode(-1); // Custom code for error
+			messageDataResponse.setMessage("Server error: " + ex.getMessage());
+		}
+		return messageDataResponse;
 	}
 
 	// fn: get detail order
